% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{commit}
\alias{commit}
\title{Commit command with qsub}
\usage{
commit(command, disp_name, num_jobs = NULL, file_sublist = NULL)
}
\arguments{
\item{command}{The command to be committed.}

\item{disp_name}{The display name of the job.}

\item{num_jobs}{Number of jobs. It only affect the messages. If not \code{NULL},
the message will add \code{"array"} and show the number of jobs.}

\item{file_subjects}{The file storing all the subjects to be processed. If
both \code{num_jobs} and \code{message} is no \code{NULL}, the message will show the file
name.}
}
\value{
The job id of the committed job.
}
\description{
Commit command with qsub
}
