% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{execute_jobs}
\alias{execute_jobs}
\title{Execute given jobs}
\usage{
execute_jobs(jobs)
}
\arguments{
\item{jobs}{A data frame of jobs to be executed.}
}
\value{
Invisible \code{NULL}. The function is used for its side effects.
}
\description{
Execute given jobs
}
