% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{extract_todo}
\alias{extract_todo}
\title{Extract todo jobs based on parsed command line arguments}
\usage{
extract_todo(jobs)
}
\arguments{
\item{jobs}{A \code{data.frame()} of the required jobs.}
}
\value{
The extracted jobs to be done.
}
\description{
Extract todo jobs based on parsed command line arguments
}
