% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_arguments}
\alias{parse_arguments}
\title{Parse command line arguments for scripts}
\usage{
parse_arguments()
}
\value{
A list with argument values. The same as those from
\code{\link[argparser:parse_args]{argparser::parse_args()}}.
}
\description{
Parse command line arguments for scripts
}
