% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_data_file_sum}
\alias{validate_data_file_sum}
\title{Validate data based on file counts}
\usage{
validate_data_file_sum(
  type,
  path = NULL,
  subject = NULL,
  part = NULL,
  check = FALSE
)
}
\arguments{
\item{type}{The type of data to be checked. Currently supported are
\code{"heudiconv"} and \code{"fmriprep"}.}

\item{path, subject}{The path or subject to check. Note only one of these two
parameters could be specified.}

\item{part}{The part of data to be checked, typically the last element of
\code{path}. If not specified, will extract from \code{path}.}

\item{check}{A logical value indicating if count of files should be checked.}
}
\value{
A character scalar. Possible values are: \code{"done"}, "\code{todo}",
"\code{incomplete}".
}
\description{
Used for generated data validation.
}
